+++ client-react/src/ApiHelpers/ServerFarmService.ts
@@ -40,7 +40,7 @@ export default class ServerFarmService {
 
   public static getTotalSitesIncludingSlotsInServerFarm = (subscriptionId: string, resourceId: string) => {
     const queryString =
-      `where type == 'microsoft.web/serverfarms' or type == 'microsoft.web/sites/slots'` +
+      `where type == 'microsoft.web/sites' or type == 'microsoft.web/sites/slots'` +
       `| where properties.serverFarmId == '${resourceId}'` +
       `| summarize count()`;
