+++ client-react/src/utils/scenario-checker/scenario-ids.ts
@@ -102,5 +102,5 @@ export class ScenarioIds {
   public static readonly showConnnectionStringFunctionInfo = 'showConnnectionStringFunctionInfo';
   public static readonly azureBlobMount = 'azureBlobMount';
   public static readonly showGeneralSettings = 'showGeneralSettings';
-  public static readonly appDensity = 'appDensity';
+  public static readonly isAppDensityEnabled = 'isAppDensityEnabled';
 }
