+++ client-react/src/utils/scenario-checker/scenario.models.ts
@@ -1,9 +1,11 @@
 import { ArmObj } from '../../models/arm-obj';
 import { Site } from '../../models/site/site';
+import { ServerFarm } from '../../models/serverFarm/serverfarm';
 
 // tslint:disable-next-line:interface-name
 export interface ScenarioCheckInput {
   site?: ArmObj<Site>;
+  serverFarm?: ArmObj<ServerFarm>;
 }
 
 export type ScenarioStatus = 'enabled' | 'disabled' | null;
